# GS-HW: Optimise readback performance for some cases