@refractionpcsx2 2023-02-13
I mean you can try it, I'm not expecting them to change too much since they only do 1 readback.  It might help if it only has to readback a smaller amount than before.