@TransparentBlue 2023-02-14
Tested with GT3 and the results are inconsistent but this seems a bit faster on night stages than the latest nightly for me (~115 fps vs ~117 fps at the starting line of SS Route 5 off the same savestate)