@refractionpcsx2 2023-02-14
> Tested with GT3 and the results are inconsistent but this seems a bit faster on night stages than the latest nightly for me (~115 fps vs ~117 fps at the starting line of SS Route 5 off the same savestate)

Thanks for testing, honestly not expecting much on the GT games, as I believe it does a single full screen read every frame, which is quite difficult to optimise for lol, so it's kind of out of scope of this PR.  Though if what it's reading hasn't been completely written to, it might be a touch faster, yeah.